+++ ChangeLog.rst
@@ -1,3 +1,10 @@
+development
+===========
+
+ * deprecated no_atexit in favor of atexit keyword parameter, 
+   https://github.com/amitu/importd/issues/1
+ 
+
 0.1.4 - 22-Oct-2012
 ===================